Added data annotations to validate forms before sending them. Any invalid data will return the user to the form with an error message.

+++ VoedselVerspillingApp/Controllers/HomeController.cs
@@ -4,6 +4,7 @@ using Domain.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using VoedselVerspillingApp.Models;
 using VoedselVerspillingApp.ViewModels;
 
@@ -46,7 +47,7 @@ public class HomeController : Controller
                 ViewBag.CanteenId = _employeeRepository.GetEmployeeByEmail(User.Identity!.Name!)!.Canteen.Id;
             }
         }
-        
+
         return View(_mealBoxRepository.GetAvailableMealBoxes());
     }
 
@@ -60,6 +61,7 @@ public class HomeController : Controller
         {
             return RedirectToAction("AvailableMealBoxes");
         }
+
         return RedirectToAction("Error");
     }
 
@@ -84,8 +86,8 @@ public class HomeController : Controller
     {
         var list = _productRepository.GetProducts();
         var listCheck = list.Select(p => new CheckBoxItem { Name = p.Name, Id = p.Id, IsChecked = false }).ToList();
-        ViewBag.Products = listCheck;
         var e = _employeeRepository.GetEmployeeByEmail(User.Identity!.Name!)!;
+        ViewBag.Products = listCheck;
         ViewBag.CanteenId = e.CanteenId;
         ViewBag.City = e.Canteen.City;
         ViewBag.ServesWarmMeals = e.Canteen.ServesWarmMeals;
@@ -99,14 +101,31 @@ public class HomeController : Controller
     [HttpPost]
     public IActionResult CreateMealBox(MealBox mealBox, MealBoxViewModel model)
     {
-        var list = model.ProductIds.Select(id => _productRepository.GetProduct(id)!).ToList();
-        foreach (var id in model.ProductIds.Where(id => _productRepository.GetProduct(id)!.IsAlcoholic))
+        if (ModelState.IsValid)
         {
-            mealBox.IsEighteen = true;
+            var list = model.ProductIds.Select(id => _productRepository.GetProduct(id)).ToList() ??
+                       new List<Product?>();
+            foreach (var id in model.ProductIds.Where(id => _productRepository.GetProduct(id)!.IsAlcoholic))
+            {
+                mealBox.IsEighteen = true;
+            }
+
+            mealBox.Products = list;
+            _mealBoxRepository.CreateMealBox(mealBox);
+            return RedirectToAction("AvailableMealBoxes");
         }
-        mealBox.Products = list;
-        _mealBoxRepository.CreateMealBox(mealBox);
-        return RedirectToAction("AvailableMealBoxes");
+
+        // Code has to be repeated from create meal box or forms won't be validated.
+        var productList = _productRepository.GetProducts();
+        var listCheck = productList.Select(p => new CheckBoxItem { Name = p.Name, Id = p.Id, IsChecked = false })
+            .ToList();
+        var e = _employeeRepository.GetEmployeeByEmail(User.Identity!.Name!)!;
+        ViewBag.Products = listCheck;
+        ViewBag.CanteenId = e.CanteenId;
+        ViewBag.City = e.Canteen.City;
+        ViewBag.ServesWarmMeals = e.Canteen.ServesWarmMeals;
+        model.Products = listCheck;
+        return View(model);
     }
 
     [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
@@ -125,16 +144,11 @@ public class HomeController : Controller
         var e = _employeeRepository.GetEmployeeByEmail(User.Identity!.Name!)!;
         ViewBag.ServesWarmMeals = e.Canteen.ServesWarmMeals;
         var listCheck = list.Select(p => new CheckBoxItem { Name = p.Name, Id = p.Id, IsChecked = false }).ToList();
-        var productIds = new List<int>();
-        foreach (Product p in box.Products)
-        {
-            productIds.Add(p.Id);
-        }
+        var productIds = box.Products.Select(p => p.Id).ToList();
         foreach (var c in listCheck.Where(c => productIds.Contains(c.Id)))
         {
             c.IsChecked = true;
         }
-
         model.Id = box.Id;
         model.Name = box.Name;
         model.PickUpTime = box.PickUpTime;
@@ -153,45 +167,62 @@ public class HomeController : Controller
     [Authorize(Roles = "Employee")]
     public IActionResult EditMealBox(MealBox mealBox, string s, MealBoxViewModel model)
     {
-        switch (s)
+        if (ModelState.IsValid)
         {
-            case "edit":
+            switch (s)
             {
-                if (mealBox.StudentId == null)
+                case "edit":
                 {
-                    var alcoholicProductsCount = model.ProductIds.Count(id => _productRepository.GetProduct(id)!.IsAlcoholic);
-                    
-                    var box = new MealBox()
+                    if (mealBox.StudentId == null)
                     {
-                        Name = model.Name,
-                        City = model.City,
-                        PickUpTime = model.PickUpTime,
-                        PickUpBy = model.PickUpBy,
-                        IsEighteen = alcoholicProductsCount > 0,
-                        Price = model.Price,
-                        MealType = model.MealType,
-                        CanteenId = model.CanteenId,
-                        Products = model.ProductIds.Select(id => _productRepository.GetProduct(id)!).ToList()
-                    };
-                    
-                    _mealBoxRepository.DeleteMealBox(_mealBoxRepository.GetMealBox(model.Id)!);
-                    _mealBoxRepository.CreateMealBox(box);
-                    return RedirectToAction("AvailableMealBoxes");
-                }
+                        var alcoholicProductsCount =
+                            model.ProductIds.Count(id => _productRepository.GetProduct(id)!.IsAlcoholic);
 
-                break;
-            }
-            case "delete":
-            {
-                if (mealBox.StudentId == null)
+                        var box = new MealBox()
+                        {
+                            Name = model.Name,
+                            City = model.City,
+                            PickUpTime = model.PickUpTime,
+                            PickUpBy = model.PickUpBy,
+                            IsEighteen = alcoholicProductsCount > 0,
+                            Price = model.Price,
+                            MealType = model.MealType,
+                            CanteenId = model.CanteenId,
+                            Products = model.ProductIds.Select(id => _productRepository.GetProduct(id)!).ToList()
+                        };
+
+                        _mealBoxRepository.DeleteMealBox(_mealBoxRepository.GetMealBox(model.Id)!);
+                        _mealBoxRepository.CreateMealBox(box);
+                        return RedirectToAction("AvailableMealBoxes");
+                    }
+
+                    break;
+                }
+                case "delete":
                 {
-                    _mealBoxRepository.DeleteMealBox(mealBox);
+                    if (mealBox.StudentId == null)
+                    {
+                        _mealBoxRepository.DeleteMealBox(mealBox);
+                    }
+
+                    break;
                 }
-
-                break;
             }
-        }
 
-        return RedirectToAction("AvailableMealBoxes");
+            return RedirectToAction("AvailableMealBoxes");
+        }
+        // Code needs to be repeated for validation to function
+        var oldBox = _mealBoxRepository.GetMealBox(mealBox.Id)!;
+        var list = _productRepository.GetProducts();
+        var e = _employeeRepository.GetEmployeeByEmail(User.Identity!.Name!)!;
+        ViewBag.ServesWarmMeals = e.Canteen.ServesWarmMeals;
+        var listCheck = list.Select(p => new CheckBoxItem { Name = p.Name, Id = p.Id, IsChecked = false }).ToList();
+        var productIds = oldBox.Products.Select(p => p.Id).ToList();
+        foreach (var c in listCheck.Where(c => productIds.Contains(c.Id)))
+        {
+            c.IsChecked = true;
+        }
+        model.Products = listCheck;
+        return View(model);
     }
 }
